own class for rollback
