+++ mt-kahypar/partition/refinement/gains/km1/km1_gain_cache.h
@@ -34,6 +34,7 @@
 #include "mt-kahypar/datastructures/hypergraph_common.h"
 #include "mt-kahypar/datastructures/array.h"
 #include "mt-kahypar/datastructures/sparse_map.h"
+#include "mt-kahypar/partition/refinement/gains/km1/km1_rollback.h"
 #include "mt-kahypar/parallel/atomic_wrapper.h"
 #include "mt-kahypar/macros.h"
 
@@ -41,7 +42,6 @@ namespace mt_kahypar {
 
 // Forward
 class DeltaKm1GainCache;
-class Km1Rollback;
 
 class Km1GainCache final : public kahypar::meta::PolicyBase {
 
@@ -368,56 +368,6 @@ class DeltaKm1GainCache {
   // ! Stores the delta of each locally touched gain cache entry
   // ! relative to the gain cache in '_phg'
   ds::DynamicFlatMap<size_t, HyperedgeWeight> _gain_cache_delta;
-};
-
-class Km1Rollback {
-
- public:
-  struct RecalculationData {
-    MoveID first_in, last_out;
-    HypernodeID remaining_pins;
-    RecalculationData() :
-      first_in(std::numeric_limits<MoveID>::max()),
-      last_out(std::numeric_limits<MoveID>::min()),
-      remaining_pins(0)
-      { }
-
-    void reset() {
-      first_in = std::numeric_limits<MoveID>::max();
-      last_out = std::numeric_limits<MoveID>::min();
-      remaining_pins = 0;
-    }
-  };
-
-  static void updateMove(const MoveID m_id,
-                         const Move& m,
-                         vec<RecalculationData>& r) {
-    r[m.to].first_in = std::min(r[m.to].first_in, m_id);
-    r[m.from].last_out = std::max(r[m.from].last_out, m_id);
-  }
-
-  static void updateNonMovedPinInBlock(const PartitionID block,
-                                       vec<RecalculationData>& r) {
-    r[block].remaining_pins++;
-  }
-
-  template<typename PartitionedHypergraph>
-  static bool hasBenefit(const PartitionedHypergraph&,
-                         const HyperedgeID,
-                         const MoveID m_id,
-                         const Move& m,
-                         vec<RecalculationData>& r) {
-    return r[m.from].last_out == m_id && r[m.from].first_in > m_id && r[m.from].remaining_pins == 0;
-  }
-
-  template<typename PartitionedHypergraph>
-  static bool hasPenalty(const PartitionedHypergraph&,
-                         const HyperedgeID,
-                         const MoveID m_id,
-                         const Move& m,
-                         vec<RecalculationData>& r) {
-    return r[m.to].first_in == m_id && r[m.to].last_out < m_id && r[m.to].remaining_pins == 0;
-  }
 };
 
 }  // namespace mt_kahypar
